+++ tests/git.test.ts
@@ -12,6 +12,7 @@ Deno.test({
         assertEquals(GenerateGitUrl("https://github.com/user/repo.test").full, "https://github.com/user/repo.test.git");
 
         assertEquals(GenerateGitUrl("gh:user/repo").full, "https://github.com/user/repo.git");
+        assertEquals(GenerateGitUrl("gh:user/repo/").full, "https://github.com/user/repo.git");
         assertEquals(GenerateGitUrl("gh:user/repo.").full, "https://github.com/user/repo.git");
         assertEquals(GenerateGitUrl("bb:user/repo").full, "https://bitbucket.org/user/repo.git");
         assertEquals(GenerateGitUrl("gl:me/code...").full, "https://gitlab.com/me/code.git");